+++ apps/server/src/lib/api/routes.rs
@@ -1,9 +1,10 @@
 use std::sync::Arc;
 
-use axum::{extract::State, Json, http::StatusCode};
+use anyhow::anyhow;
+use axum::{extract::{Path, State}, http::StatusCode, Json};
 use serde::{Deserialize, Serialize};
 
-use crate::{game::Category, net::connection::{HostToken, RoomCode}, AppState, Room};
+use crate::{api::handlers::RoomParams, game::Category, net::connection::{HostToken, RoomCode}, AppState, Room};
 
 #[tracing::instrument(skip(state, body))]
 pub async fn create_room(
@@ -49,5 +50,50 @@ pub struct CreateRoomResponse {
 #[derive(Deserialize)]
 pub struct CreateRoomRequest {
     categories: Option<Vec<Category>>,
+}
+
+#[tracing::instrument(skip(state), fields(room_code = %rp.code))]
+pub async fn cpr_handler(
+    State(state): State<Arc<AppState>>,
+    Path(rp @ RoomParams { .. }): Path<RoomParams>,
+) -> String {
+    let code = rp.code;
+    let res = {
+        let mut room_map = state.room_map.lock().await;
+        let room_res = room_map
+            .get_mut(&code)
+            .ok_or_else(|| anyhow!("Room {} does not exist", code));
+        let mut failures = 0_u32;
+        match room_res {
+            Err(e) => Err(e),
+            Ok(room) => {
+                for entry in &mut room.players {
+                    match entry.heartbeat().await {
+                        Ok(()) => {}
+                        Err(e) => {
+                            tracing::warn!(
+                                player_id = entry.player.pid,
+                                error = %e,
+                                "Heartbeat failed"
+                            );
+                            failures += 1;
+                        }
+                    }
+                }
+                Ok(format!(
+                    "Ok, requested {} heartbeats, {} failed immediately",
+                    room.players.len(),
+                    failures
+                ))
+            }
+        }
+    };
+    match res {
+        Ok(s) => s,
+        Err(e) => {
+            tracing::error!(error = %e, "CPR handler failed");
+            format!("Err, {e}")
+        }
+    }
 }
