+++ apps/server/src/lib/net.rs
@@ -1,1 +1,2 @@
 pub mod connection;
+pub mod ws;
